src/requirements.txt
Django==4.0.6
Pillow==8.3.2
psycopg2-binary==2.9.3
gunicorn==20.1.0
django-allauth==0.51.0
django-resized==1.0.2
